requirements.txt

#base ---------------
nats-py
numpy ==1.22.0
Pillow==9.1.1
opencv-python==4.5.5.64
nanoid
torch==1.8.1  
torchvision==0.9.1
pandas>=1.1.4
seaborn>=0.11.0
requests>=2.23.0
iopath
fvcore
streamlit
pyngrok
cmake
face-recognition
nanoid
cython 
chardet
rq
nest-asyncio
statsmodels
cvxpy
pickledb
h5py
six
tb-nightly
future
yacs
gdown
flake8
yapf
easydict
lap
filterpy
GitPython
#git+https://github.com/samson-wang/cython_bbox.git@9badb346a9222c98f828ba45c63fe3b7f2790ea2



